\name{ZD}
\alias{ZD}
\docType{data}
\title{
  Zone Diameter Distributions of Microorganisms Without and With Resistance Mechanisms.
}
\description{
  
  The data give inhibition zone diameter distributions for individual
  organisms and antimicrobial agents in tables and histograms. The
  distributions are based on collated data from an increasing number of
  sources worldwide. The data are representative of results obtained
  with the EUCAST disk diffusion method (launched in 2009).
  
}
\usage{data("ZD")}
\format{
  A data frame with 1476 observations on the following 51 variables.
  \describe{
    \item{\code{Antimicrobial}}{a factor}
    \item{\code{Bacterium}}{a factor}
    \item{\code{DiskContent}}{a numeric vector}
    \item{\code{Z6}}{a numeric vector}
    \item{\code{Z7}}{a numeric vector}
    \item{\code{Z8}}{a numeric vector}
    \item{\code{Z9}}{a numeric vector}
    \item{\code{Z10}}{a numeric vector}
    \item{\code{Z11}}{a numeric vector}
    \item{\code{Z12}}{a numeric vector}
    \item{\code{Z13}}{a numeric vector}
    \item{\code{Z14}}{a numeric vector}
    \item{\code{Z15}}{a numeric vector}
    \item{\code{Z16}}{a numeric vector}
    \item{\code{Z17}}{a numeric vector}
    \item{\code{Z18}}{a numeric vector}
    \item{\code{Z19}}{a numeric vector}
    \item{\code{Z20}}{a numeric vector}
    \item{\code{Z21}}{a numeric vector}
    \item{\code{Z22}}{a numeric vector}
    \item{\code{Z23}}{a numeric vector}
    \item{\code{Z24}}{a numeric vector}
    \item{\code{Z25}}{a numeric vector}
    \item{\code{Z26}}{a numeric vector}
    \item{\code{Z27}}{a numeric vector}
    \item{\code{Z28}}{a numeric vector}
    \item{\code{Z29}}{a numeric vector}
    \item{\code{Z30}}{a numeric vector}
    \item{\code{Z31}}{a numeric vector}
    \item{\code{Z32}}{a numeric vector}
    \item{\code{Z33}}{a numeric vector}
    \item{\code{Z34}}{a numeric vector}
    \item{\code{Z35}}{a numeric vector}
    \item{\code{Z36}}{a numeric vector}
    \item{\code{Z37}}{a numeric vector}
    \item{\code{Z38}}{a numeric vector}
    \item{\code{Z39}}{a numeric vector}
    \item{\code{Z40}}{a numeric vector}
    \item{\code{Z41}}{a numeric vector}
    \item{\code{Z42}}{a numeric vector}
    \item{\code{Z43}}{a numeric vector}
    \item{\code{Z44}}{a numeric vector}
    \item{\code{Z45}}{a numeric vector}
    \item{\code{Z46}}{a numeric vector}
    \item{\code{Z47}}{a numeric vector}
    \item{\code{Z48}}{a numeric vector}
    \item{\code{Z49}}{a numeric vector}
    \item{\code{Z50}}{a numeric vector}
    \item{\code{ECOFF}}{a numeric vector}
    \item{\code{Distributions}}{a numeric vector}
    \item{\code{Observations}}{a numeric vector}
  }
}
\source{
  EUCAST database for antimicrobial wild type distributions of microorganisms.
  \url{https://mic.eucast.org/Eucast2/}
}
\examples{
data("ZD")
}
\keyword{datasets}
